moved around a substantial amount of logic - WIP
