+++ 2025_NoName/src/main/java/frc/robot/subsystems/drive/CommandSwerveDrivetrain.java
@@ -42,6 +42,7 @@ import java.util.function.Supplier;
  * be used in command-based projects.
  */
 public class CommandSwerveDrivetrain extends TunerSwerveDrivetrain implements Subsystem {
+  private final Telemetry telemetry = new Telemetry(TunerConstants.kSpeedAt12Volts.in(MetersPerSecond));
   CommandXboxController driver;
   private double MaxSpeed =
       TunerConstants.kSpeedAt12Volts.in(MetersPerSecond); // kSpeedAt12Volts desired top speed
@@ -192,76 +193,10 @@ public class CommandSwerveDrivetrain extends TunerSwerveDrivetrain implements Su
 
     choreoThetaPID.enableContinuousInput(-Math.PI, Math.PI);
 
+    this.registerTelemetry(telemetry::telemeterize);
+
     resetDriveToPoint();
   }
-
-  // /**
-  //  * Constructs a CTRE SwerveDrivetrain using the specified constants.
-  //  *
-  //  * <p>This constructs the underlying hardware devices, so users should not construct the
-  // devices
-  //  * themselves. If they need the devices, they can access them through getters in the classes.
-  //  *
-  //  * @param drivetrainConstants Drivetrain-wide constants for the swerve drive
-  //  * @param odometryUpdateFrequency The frequency to run the odometry loop. If unspecified or set
-  // to
-  //  *     0 Hz, this is 250 Hz on CAN FD, and 100 Hz on CAN 2.0.
-  //  * @param modules Constants for each specific module
-  //  */
-  // public CommandSwerveDrivetrain(
-  //     SwerveDrivetrainConstants drivetrainConstants,
-  //     double odometryUpdateFrequency,
-  //     SwerveModuleConstants<?, ?, ?>... modules) {
-  //   super(drivetrainConstants, odometryUpdateFrequency, modules);
-  //   if (Utils.isSimulation()) {
-  //     startSimThread();
-  //   }
-  //   driveAtAngle.HeadingController = new PhoenixPIDController(5, 0, 0);
-  //   driveAtAngle.HeadingController.enableContinuousInput(-Math.PI, Math.PI);
-  //   choreoThetaPID.enableContinuousInput(-Math.PI, Math.PI);
-  //   driveToPointController.setTolerance(DRIVE_TO_POINT_TRANSLATION_ERROR_TOLERANCE);
-  //   resetDriveToPoint();
-  // }
-
-  // /**
-  //  * Constructs a CTRE SwerveDrivetrain using the specified constants.
-  //  *
-  //  * <p>This constructs the underlying hardware devices, so users should not construct the
-  // devices
-  //  * themselves. If they need the devices, they can access them through getters in the classes.
-  //  *
-  //  * @param drivetrainConstants Drivetrain-wide constants for the swerve drive
-  //  * @param odometryUpdateFrequency The frequency to run the odometry loop. If unspecified or set
-  // to
-  //  *     0 Hz, this is 250 Hz on CAN FD, and 100 Hz on CAN 2.0.
-  //  * @param odometryStandardDeviation The standard deviation for odometry calculation in the form
-  //  *     [x, y, theta]ᵀ, with units in meters and radians
-  //  * @param visionStandardDeviation The standard deviation for vision calculation in the form [x,
-  // y,
-  //  *     theta]ᵀ, with units in meters and radians
-  //  * @param modules Constants for each specific module
-  //  */
-  // public CommandSwerveDrivetrain(
-  //     SwerveDrivetrainConstants drivetrainConstants,
-  //     double odometryUpdateFrequency,
-  //     Matrix<N3, N1> odometryStandardDeviation,
-  //     Matrix<N3, N1> visionStandardDeviation,
-  //     SwerveModuleConstants<?, ?, ?>... modules) {
-  //   super(
-  //       drivetrainConstants,
-  //       odometryUpdateFrequency,
-  //       odometryStandardDeviation,
-  //       visionStandardDeviation,
-  //       modules);
-  //   if (Utils.isSimulation()) {
-  //     startSimThread();
-  //   }
-  //   driveAtAngle.HeadingController = new PhoenixPIDController(5, 0, 0);
-  //   driveAtAngle.HeadingController.enableContinuousInput(-Math.PI, Math.PI);
-  //   choreoThetaPID.enableContinuousInput(-Math.PI, Math.PI);
-  //   driveToPointController.setTolerance(DRIVE_TO_POINT_TRANSLATION_ERROR_TOLERANCE);
-  //   resetDriveToPoint();
-  // }
 
   /**
    * Returns a command that applies the specified control request to this swerve drivetrain.
