+++ 2025_NoName/src/main/java/frc/robot/util/AutoAlignPoseGenerator.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.util;
 
 import static frc.robot.FieldConstants.*;
